# REST APIs Design/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  description: REST APIs for the interaction with the Film Manager service.
  termsOfService: 'http://swagger.io/terms/'
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
  version: 1.0.0
...
servers:
  - url: 'https://todo.swagger.io/v2'
  - url: 'http://todo.swagger.io/v2'
...
paths:
  '/api/films':
    post:
      tags:
        - films
      security:
        - cookieAuth: []
      summary: Create a new film
      description: A new film is created by the authenticated user.
      operationId: createFilm
      requestBody:
        description: Representation of the film to be created
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
            example:
              title: Jujutsu Kaisen
              private: false
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Film'
              example:
                id: 11
                title: Jujutsu Kaisen 0
                owner: 1
                private: false
                self: /api/films/public/11
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/JsonSchemaValidationError'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/api/films/private': 
    get:
      tags:
        - films/private
      security:
        - cookieAuth: []
      summary: Retrieve the private films of the logged-in user 
      description: >-
        The private films of the logged-in user are retrieved. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPrivateFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
          example : 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              example:
                totalPages: 1
                currentPage: 1
                totalItems: 1
                films: [{
                  id: 1,
                  title: Your Name,
                  owner: 1,
                  private: true,
                  watchDate: 2021-10-03,
                  rating: 9,
                  favorite: true,
                  self: /api/films/private/1
                }]
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/api/films/private/{filmId}':
    get:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Retrieve a private film
      description: The private film with ID filmId is retrieved. This operation can be performed on the film if the user who performs the operation is the film's owner.
      operationId: getSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to retrieve
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Film'
              example:
                id: 1,
                title: Your Name,
                owner: 1,
                private: true,
                watchDate: 2021-10-03,
                rating: 9,
                favorite: true,
                self: /api/films/private/1
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '403':
          description: Forbidden
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The user is not the owner of the film.'
                }]
        '404':
          description: Not Found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The film does not exist.'
                }]
        '500':
          description: Internal Server Error
          content:  
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Update a private film
      description: The private film with ID filmId is updated. This operation does not allow changing its visibility. This operation can be performed only by the owner.
      operationId: updateSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to update
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        description: The updated film object that needs to replace the old object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Film'
            example:
              id: 1,
              title: Your Name,
              owner: 1,
              private: true,
              watchDate: 2021-10-09,
              rating: 9,
              favorite: true
        required: true
      responses:
        '204':
          description: No Content
          content: {}
        '400':
          description: Bad Request
          content: 
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/JsonSchemaValidationError'
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '403':
          description: Forbidden
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The user is not the owner of the film.'
                }]
        '409':
          description: Conflict (attempt to make unallowed change, e.g., visibility)
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The visibility of the film cannot be changed.'
              }]
        '404':
          description: Not Found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The film does not exist.'
                }]
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - 'films/private/{filmId}'
      security:
        - cookieAuth: []
      summary: Delete a private film
      description: The private film with ID filmId is deleted. This operation can only be performed by the owner.
      operationId: deleteSinglePrivateFilm
      parameters:
        - name: filmId
          in: path
          description: ID of the film to delete
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '204':
          description: Successful operation - No Content
          content: {}
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The user is not the owner of the film.'
              }]
        '404':
          description: Not Found
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                errors: [{
                  param: 'Server',
                  msg: 'The film does not exist.'
              }]
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/api/films/public': 
    get:
      tags:
        - films/public
      summary: Retrieve the public films
      description: >-
        The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
      operationId: getPublicFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
          example: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              example:
                totalPages: 1
                currentPage: 1
                totalItems: 1
                films: [{
                  id: 1,
                  title: Your Name,
                  owner: 1,
                  private: false,
                  watchDate: 2021-10-03,
                  rating: 9,
                  favorite: true,
                  self: /api/films/public/1
                }]
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 
  '/api/films/public/invited': 
    get:
      tags:
        - films/public/invited
      security:
        - cookieAuth: []
      summary: Retrieve the public films that the logged-in user has been invited to review
      description: >-
        The public films that the logged-in user has been invited to review are retrieved. A pagination mechanism is implemented to limit the size of messages.
      operationId: getInvitedFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
          description: 'The id of the requested page (if absent, the first page is returned)'
          example: 1
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalPages:
                    type: integer
                  currentPage:
                    type: integer
                  totalItems:
                    type: integer
                  films:
                    type: array
                    items:
                      $ref: '#/components/schemas/Film'
                  next:
                    type: string
                    format: uri
              example:
                totalPages: 1
                currentPage: 1
                totalItems: 1
                films: [{
                  id: 1,
                  title: Your Name,
                  owner: 1,
                  private: false,
                  watchDate: 2021-10-03,
                  rating: 9,
                  favorite: true,
                  self: /api/films/public/1
                }]
        '401':
          description: Unauthorized
          content: 
            application/json:
              schema:
                type: object
                properties: 
                  error:
                    type: string
                    enum: ['Not authorized']
                additionalProperties: false
                required: 
                  - error
              example:
                error: 'Not authorized'
        '500':
          description: Internal Server Error
          content: 
            application/json:
              schema:
                $ref: '#/components/schemas/Error' 
...
components:
  schemas:
...
    Film:
      description: schema for Film data structures
      type: object
      additionalProperties: false
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON data structure
            describing the film
        id:
          type: integer
          description: unique identifier for the film data structure
        title:
          type: string
          description: textual title of the film
        owner:
          type: integer
          description: unique identifier of the the film data structure owner
        private:
          type: boolean
          description: true if only the owner of the film data structure can access it, false otherwise
          default: 'true'
        watchDate:
          type: string
          format: date
          description: the date when the film was watched by the owner
        rating:
          type: integer
          description: the rating the owner has given to the film
          minimum: 1
          maximum: 10
        favorite:
          type: boolean
          description: 'true if the film is among the favourite
            ones of the owner, false otherwise '
          default: 'false'
        self: 
          type: string
          format: uri
      required:
      - title
      - owner
      - private
...
    Error:
      description: schema for Error data structures
      type: object
      properties:
        errors:
          type: array
          minItems: 1
          maxItems: 1
          items:
            type: object
            properties:
              param:
                type: string
                enum: ['Server']
              msg:
                description: 'Explanation of the error'
                type: string
            additionalProperties: false
            required:
              - param
              - msg
      additionalProperties: false
      required:
      - errors
    JsonSchemaValidationError:
      description: schema for JsonSchemaValidationError data structures
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          enum: ['JsonSchemaValidationError']
        validationErrors:
          type: object
          properties:
            body:
              type: array
              items:
                type: object
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid